android: add confirmation dialog on app close
The back button triggers an application close.  So ask the user first,
if he wants to leave the current game.  We might add an option to save
later.  In the meanwhile the user can switch applications by holding
home down for a bit.

This patch also contains some minor refactorings in the MainLoop.