Working serialization code.
Game serialization works.  You can serialize an AWT game by pressing 0.
This will save the game's state into the file "savegame" in the current
directory.  By invoking the MainLoop with the filename as the first
parameter you can deserialize it.

Known glitch: The paused state is not properly saved.  If the game is
saved paused the overlay will be drawn but the enemies will walk again.